Issue #708 - it is now possible to build edex rpms in the new baseline repository.
Former-commit-id: f79134499eeb7f88bd738f27fa0a4babd392c3ed [formerly 0932e1173258d916a6a1caae5c943903d830c2af]
Former-commit-id: 973ceff9a72916aade39e7a0a12daa3a337675fe

+++ rpms/awips2.edex/Installer.edex-base/component.spec
@@ -54,7 +54,7 @@ copyApplicableDeltas ${RPM_BUILD_ROOT} %{_component_name} \
 #---------------------------------------------------------------------------#
 
 %install
-DEPLOY_SCRIPT="build.edex/deploy-install.xml"
+DEPLOY_SCRIPT="build.edex/deploy-common/deploy-esb.xml"
 
 # Deploy Edex To Our Temporary Build Directory.
 
@@ -67,14 +67,18 @@ if [ ! "${RC}" = "0" ]; then
    exit 1
 fi
 
-ANT_EXE=`rpm -q --queryformat '%{INSTALLPREFIX}\n' awips2-ant`
-ANT_EXE="${ANT_EXE}/bin/ant"
+ANT_EXE="/awips2/ant/bin/ant"
 
 ${ANT_EXE} -file ${WORKSPACE_DIR}/${DEPLOY_SCRIPT} \
-   -Dinstall.dir=${RPM_BUILD_ROOT}/awips2/edex \
-   -Dinstaller=true -Dlocal.build=false \
-   -Dcomponent.to.deploy=edex-base
-   
+   -Ddeploy.data=true -Ddeploy.web=true \
+   -Desb.overwrite=true \
+   -Desb.directory=${WORKSPACE_DIR}/build.edex/esb \
+   -Dedex.root.directory=${RPM_BUILD_ROOT}/awips2/edex \
+   -Dbasedir=${WORKSPACE_DIR}/build.edex \
+   -Dbasedirectories=${WORKSPACE_DIR}
+if [ $? -ne 0 ]; then
+   exit 1
+fi
 # Remove the directory with the shapefiles (if it exists)
 if [ -d ${RPM_BUILD_ROOT}/awips2/edex/data/utility/edex_static/base/shapefiles ]; then
    rm -rf ${RPM_BUILD_ROOT}/awips2/edex/data/utility/edex_static/base/shapefiles
